# docs/api/openapi.yaml
components:
  schemas:
...
    AuditLogDetailResponse:
      description: Detailed audit log with request/response data.
      properties:
        action:
          title: Action
          type: string
        duration_ms:
          anyOf:
          - type: string
          - type: 'null'
          title: Duration Ms
        endpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Endpoint
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        extra_data:
          anyOf:
          - type: object
          - type: 'null'
          title: Extra Data
        id:
          title: Id
          type: string
        ip_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip Address
        method:
          anyOf:
          - type: string
          - type: 'null'
          title: Method
        request_data:
          anyOf:
          - type: object
          - type: 'null'
          title: Request Data
        resource_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Id
        resource_type:
          title: Resource Type
          type: string
        response_data:
          anyOf:
          - type: object
          - type: 'null'
          title: Response Data
        status:
          title: Status
          type: string
        status_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Code
        timestamp:
          format: date-time
          title: Timestamp
          type: string
        user_agent:
          anyOf:
          - type: string
          - type: 'null'
          title: User Agent
        user_email:
          anyOf:
          - type: string
          - type: 'null'
          title: User Email
        user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      required:
      - id
      - timestamp
      - user_id
      - user_email
      - action
      - resource_type
      - resource_id
      - method
      - endpoint
      - ip_address
      - status
      - status_code
      - error_message
      - duration_ms
      - request_data
      - response_data
      - extra_data
      - user_agent
      title: AuditLogDetailResponse
      type: object
    AuditLogResponse:
      description: Schema for audit log response.
      properties:
        action:
          title: Action
          type: string
        duration_ms:
          anyOf:
          - type: string
          - type: 'null'
          title: Duration Ms
        endpoint:
          anyOf:
          - type: string
          - type: 'null'
          title: Endpoint
        error_message:
          anyOf:
          - type: string
          - type: 'null'
          title: Error Message
        id:
          title: Id
          type: string
        ip_address:
          anyOf:
          - type: string
          - type: 'null'
          title: Ip Address
        method:
          anyOf:
          - type: string
          - type: 'null'
          title: Method
        resource_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Id
        resource_type:
          title: Resource Type
          type: string
        status:
          title: Status
          type: string
        status_code:
          anyOf:
          - type: string
          - type: 'null'
          title: Status Code
        timestamp:
          format: date-time
          title: Timestamp
          type: string
        user_email:
          anyOf:
          - type: string
          - type: 'null'
          title: User Email
        user_id:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      required:
      - id
      - timestamp
      - user_id
      - user_email
      - action
      - resource_type
      - resource_id
      - method
      - endpoint
      - ip_address
      - status
      - status_code
      - error_message
      - duration_ms
      title: AuditLogResponse
      type: object
    AuditStatsResponse:
      description: Audit statistics.
      properties:
        avg_duration_ms:
          anyOf:
          - type: number
          - type: 'null'
          title: Avg Duration Ms
        events_by_action:
          title: Events By Action
          type: object
        events_by_status:
          title: Events By Status
          type: object
        events_by_user:
          title: Events By User
          type: object
        failed_events:
          title: Failed Events
          type: integer
        security_events:
          title: Security Events
          type: integer
        total_events:
          title: Total Events
          type: integer
      required:
      - total_events
      - events_by_action
      - events_by_user
      - events_by_status
      - failed_events
      - security_events
      - avg_duration_ms
      title: AuditStatsResponse
      type: object
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          title: Detail
          type: array
      title: HTTPValidationError
      type: object
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          title: Location
          type: array
        msg:
          title: Message
          type: string
        type:
          title: Error Type
          type: string
      required:
      - loc
      - msg
      - type
      title: ValidationError
      type: object
...
  securitySchemes:
    APIKeyHeader:
      in: header
      name: X-API-Key
      type: apiKey
info:
  description: BacklinkContent Engine - AI-powered content generation for SEO
  title: BACOWR API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /api/v1/audit/failed-actions:
    get:
      description: "Get recent failed actions (admin only).\n\nArgs:\n    hours: Look\
        \ back N hours\n    limit: Max results\n    current_user: Current admin user\n\
        \    db: Database session\n\nReturns:\n    List of failed action audit logs"
      operationId: get_failed_actions_api_v1_audit_failed_actions_get
      parameters:
      - in: query
        name: hours
        required: false
        schema:
          default: 24
          maximum: 720
          minimum: 1
          title: Hours
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AuditLogResponse'
                title: Response Get Failed Actions Api V1 Audit Failed Actions Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyHeader: []
      summary: Get Failed Actions
      tags:
      - audit
  /api/v1/audit/logs:
    get:
      description: "List audit logs (admin only).\n\nArgs:\n    limit: Max results\n\
        \    offset: Offset for pagination\n    action: Filter by action\n    user_id:\
        \ Filter by user\n    resource_type: Filter by resource type\n    status_filter:\
        \ Filter by status\n    hours: Look back N hours\n    current_user: Current\
        \ admin user\n    db: Database session\n\nReturns:\n    List of audit logs"
      operationId: list_audit_logs_api_v1_audit_logs_get
      parameters:
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: action
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      - in: query
        name: user_id
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: User Id
      - in: query
        name: resource_type
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Resource Type
      - in: query
        name: status
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Status
      - in: query
        name: hours
        required: false
        schema:
          anyOf:
          - maximum: 720
            minimum: 1
            type: integer
          - type: 'null'
          title: Hours
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AuditLogResponse'
                title: Response List Audit Logs Api V1 Audit Logs Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyHeader: []
      summary: List Audit Logs
      tags:
      - audit
  /api/v1/audit/logs/{log_id}:
    get:
      description: "Get detailed audit log entry (admin only).\n\nArgs:\n    log_id:\
        \ Audit log ID\n    current_user: Current admin user\n    db: Database session\n\
        \nReturns:\n    Detailed audit log\n\nRaises:\n    404: If log not found"
      operationId: get_audit_log_api_v1_audit_logs__log_id__get
      parameters:
      - in: path
        name: log_id
        required: true
        schema:
          title: Log Id
          type: string
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditLogDetailResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyHeader: []
      summary: Get Audit Log
      tags:
      - audit
  /api/v1/audit/resources/{resource_type}/{resource_id}/history:
    get:
      description: "Get audit history for specific resource (admin only).\n\nArgs:\n\
        \    resource_type: Resource type (job, batch, etc.)\n    resource_id: Resource\
        \ ID\n    limit: Max results\n    current_user: Current admin user\n    db:\
        \ Database session\n\nReturns:\n    List of resource audit logs"
      operationId: get_resource_history_api_v1_audit_resources__resource_type___resource_id__history_get
      parameters:
      - in: path
        name: resource_type
        required: true
        schema:
          title: Resource Type
          type: string
      - in: path
        name: resource_id
        required: true
        schema:
          title: Resource Id
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AuditLogResponse'
                title: Response Get Resource History Api V1 Audit Resources  Resource
                  Type   Resource Id  History Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyHeader: []
      summary: Get Resource History
      tags:
      - audit
  /api/v1/audit/security-events:
    get:
      description: "Get recent security events (admin only).\n\nArgs:\n    hours:\
        \ Look back N hours\n    limit: Max results\n    current_user: Current admin\
        \ user\n    db: Database session\n\nReturns:\n    List of security event audit\
        \ logs"
      operationId: get_security_events_api_v1_audit_security_events_get
      parameters:
      - in: query
        name: hours
        required: false
        schema:
          default: 24
          maximum: 720
          minimum: 1
          title: Hours
          type: integer
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AuditLogResponse'
                title: Response Get Security Events Api V1 Audit Security Events Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyHeader: []
      summary: Get Security Events
      tags:
      - audit
  /api/v1/audit/stats:
    get:
      description: "Get audit statistics (admin only).\n\nArgs:\n    hours: Look back\
        \ N hours\n    current_user: Current admin user\n    db: Database session\n\
        \nReturns:\n    Audit statistics"
      operationId: get_audit_stats_api_v1_audit_stats_get
      parameters:
      - in: query
        name: hours
        required: false
        schema:
          default: 24
          maximum: 720
          minimum: 1
          title: Hours
          type: integer
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditStatsResponse'
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyHeader: []
      summary: Get Audit Stats
      tags:
      - audit
  /api/v1/audit/users/{user_id}/activity:
    get:
      description: "Get audit trail for specific user (admin only).\n\nArgs:\n   \
        \ user_id: User ID\n    limit: Max results\n    offset: Offset for pagination\n\
        \    action: Filter by action\n    hours: Look back N hours\n    current_user:\
        \ Current admin user\n    db: Database session\n\nReturns:\n    List of user's\
        \ audit logs"
      operationId: get_user_activity_api_v1_audit_users__user_id__activity_get
      parameters:
      - in: path
        name: user_id
        required: true
        schema:
          title: User Id
          type: string
      - in: query
        name: limit
        required: false
        schema:
          default: 100
          maximum: 1000
          minimum: 1
          title: Limit
          type: integer
      - in: query
        name: offset
        required: false
        schema:
          default: 0
          minimum: 0
          title: Offset
          type: integer
      - in: query
        name: action
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Action
      - in: query
        name: hours
        required: false
        schema:
          anyOf:
          - maximum: 720
            minimum: 1
            type: integer
          - type: 'null'
          title: Hours
      responses:
        '200':
          content:
            application/json:
              schema:
                items:
                  $ref: '#/components/schemas/AuditLogResponse'
                title: Response Get User Activity Api V1 Audit Users  User Id  Activity
                  Get
                type: array
          description: Successful Response
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
          description: Validation Error
      security:
      - APIKeyHeader: []
      summary: Get User Activity
      tags:
      - audit
...